+++ sources/server/database/common/versioning/Version_3_0_1.h
@@ -41,7 +41,7 @@ namespace database
             /// \brief     Create the database (when tables are not found)
             ///\param [in] pRequester : database requester object
             //-----------------------------------
-            static void UpdateFrom3_0_0(const boost::shared_ptr<IDatabaseRequester>& pRequester);
+            void UpdateFrom3_0_0(const boost::shared_ptr<IDatabaseRequester>& pRequester) const;
          };
       } //namespace versioning
    } //namespace common