+++ sources/server/database/common/versioning/Version_4_0_1.cpp
@@ -49,7 +49,7 @@ namespace database
          ///\param [in] pRequester : database requester object
          ///\throw      CVersionException if create database failed
          //-----------------------------------
-         void CVersion_4_0_1::UpdateFrom4_0_0(const boost::shared_ptr<IDatabaseRequester>& pRequester)
+         void CVersion_4_0_1::UpdateFrom4_0_0(const boost::shared_ptr<IDatabaseRequester>& pRequester) const
          {
             try
             {