+++ sources/server/web/rest/service/Plugin.cpp
@@ -58,12 +58,12 @@ namespace web
          }
 
 
-         shared::CDataContainer CPlugin::transactionalMethod(CRestDispatcher::CRestMethodHandler realMethod,
-                                                             const std::vector<std::string>& parameters,
-                                                             const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::transactionalMethod(CRestDispatcher::CRestMethodHandler realMethod,
+                                                                                                  const std::vector<std::string>& parameters,
+                                                                                                  const std::string& requestContent) const
          {
             auto pTransactionalEngine = m_dataProvider->getTransactionalEngine();
-            shared::CDataContainer result;
+            boost::shared_ptr<shared::serialization::IDataSerializable> result;
             try
             {
                if (pTransactionalEngine)
@@ -81,7 +81,7 @@ namespace web
 
             if (pTransactionalEngine)
             {
-               if (CResult::isSuccess(result))
+               if (CResult::isSuccess(*boost::dynamic_pointer_cast<shared::CDataContainer>(result)))
                   pTransactionalEngine->transactionCommit();
                else
                   pTransactionalEngine->transactionRollback();
@@ -89,8 +89,8 @@ namespace web
             return result;
          }
 
-         shared::CDataContainer CPlugin::getOnePlugin(const std::vector<std::string>& parameters,
-                                                      const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getOnePlugin(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -115,8 +115,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::getAllPluginsInstance(const std::vector<std::string>& parameters,
-                                                               const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getAllPluginsInstance(const std::vector<std::string>& parameters,
+                                                                                                    const std::string& requestContent) const
          {
             auto hwList = m_pluginManager->getInstanceList();
             shared::CDataContainer t;
@@ -124,8 +124,8 @@ namespace web
             return CResult::GenerateSuccess(t);
          }
 
-         shared::CDataContainer CPlugin::getAllPluginsInstanceForManualDeviceCreation(const std::vector<std::string>& parameters,
-                                                                                      const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getAllPluginsInstanceForManualDeviceCreation(const std::vector<std::string>& parameters,
+                                                                                                                           const std::string& requestContent) const
          {
             std::vector<boost::shared_ptr<database::entities::CPlugin>> result;
 
@@ -154,8 +154,8 @@ namespace web
             return CResult::GenerateSuccess(t);
          }
 
-         shared::CDataContainer CPlugin::getAllAvailablePlugins(const std::vector<std::string>& parameters,
-                                                                const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getAllAvailablePlugins(const std::vector<std::string>& parameters,
+                                                                                                     const std::string& requestContent) const
          {
             try
             {
@@ -180,8 +180,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::getAllAvailablePluginsWithPackage(const std::vector<std::string>& parameters,
-                                                                           const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getAllAvailablePluginsWithPackage(const std::vector<std::string>& parameters,
+                                                                                                                const std::string& requestContent) const
          {
             try
             {
@@ -210,8 +210,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::createPlugin(const std::vector<std::string>& parameters,
-                                                      const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::createPlugin(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -232,8 +232,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::updatePlugin(const std::vector<std::string>& parameters,
-                                                      const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::updatePlugin(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -255,8 +255,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::sendExtraQuery(const std::vector<std::string>& parameters,
-                                                        const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::sendExtraQuery(const std::vector<std::string>& parameters,
+                                                                                             const std::string& requestContent) const
          {
             try
             {
@@ -291,8 +291,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::deletePlugin(const std::vector<std::string>& parameters,
-                                                      const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::deletePlugin(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -315,8 +315,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::deleteAllPlugins(const std::vector<std::string>& parameters,
-                                                          const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::deleteAllPlugins(const std::vector<std::string>& parameters,
+                                                                                               const std::string& requestContent) const
          {
             auto hwList = m_pluginManager->getInstanceList();
 
@@ -326,7 +326,7 @@ namespace web
             return CResult::GenerateSuccess();
          }
 
-         shared::CDataContainer CPlugin::getInstanceState(const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getInstanceState(const std::vector<std::string>& parameters, const std::string& requestContent) const
          {
             try
             {
@@ -353,7 +353,7 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::getInstanceRunning(const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getInstanceRunning(const std::vector<std::string>& parameters, const std::string& requestContent) const
          {
             try
             {
@@ -378,8 +378,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::getPluginDevices(const std::vector<std::string>& parameters,
-                                                          const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getPluginDevices(const std::vector<std::string>& parameters,
+                                                                                               const std::string& requestContent) const
          {
             try
             {
@@ -408,8 +408,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::startInstance(const std::vector<std::string>& parameters,
-                                                       const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::startInstance(const std::vector<std::string>& parameters,
+                                                                                            const std::string& requestContent) const
          {
             try
             {
@@ -444,8 +444,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::stopInstance(const std::vector<std::string>& parameters,
-                                                      const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::stopInstance(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -473,8 +473,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::getInstanceLog(const std::vector<std::string>& parameters,
-                                                        const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getInstanceLog(const std::vector<std::string>& parameters,
+                                                                                             const std::string& requestContent) const
          {
             try
             {
@@ -519,8 +519,8 @@ namespace web
             return std::string("manuallyCreatedDevice_") + std::to_string(lastNumber + 1);
          }
 
-         shared::CDataContainer CPlugin::createDevice(const std::vector<std::string>& parameters,
-                                                      const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::createDevice(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -610,8 +610,8 @@ namespace web
             }
          }
 
-         shared::CDataContainer CPlugin::getBinding(const std::vector<std::string>& parameters,
-                                                    const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CPlugin::getBinding(const std::vector<std::string>& parameters,
+                                                                                         const std::string& requestContent) const
          {
             try
             {