+++ tests/user/database.py
@@ -5,8 +5,14 @@ import os.path
 def new():
    print 'Use new database'
    resources.deleteFile(yadomsServer.databasePath())
+   resources.deleteFile(yadomsServer.databasePath() + "-journal")
 
          
 def deploy(databaseName):
-   print 'Deploy database ', databaseName
-   resources.deployFile(os.path.join("resources", "databases", databaseName + ".db3"), yadomsServer.databasePath())
+   deployFromPath(os.path.join("resources", "databases", databaseName + ".db3"))
+
+         
+def deployFromPath(databaseFullPath):
+   print 'Deploy database from ', databaseFullPath
+   resources.deleteFile(yadomsServer.databasePath() + "-journal")
+   resources.deployFile(databaseFullPath, yadomsServer.databasePath())