+++ sources/server/changelog.md
@@ -2,9 +2,14 @@
 
 #### New features and fixes
 
-* Web Client : Optimize REST request handling at the startup of the web client
-* Web Client : fix rolling title, when change the title of the widget
-* Web client : widgets configured with missing device/keyword(s) are automatically disabled
+##### Yadoms Core
+* Fix the reset to default function on system configuration page
+* Refactor the system configuration management
+
+##### Web Client
+* Optimize REST request handling at the startup of the web client
+* Fix rolling title, when change the title of the widget
+* Widgets configured with missing device/keyword(s) are automatically disabled
 
 ### 2.0.2