+++ sources/www/widgets/clock/changelog.md
@@ -1,2 +1,5 @@
+### 2.0.3
+* Fix display with v2.1.0.
+
 ### 2.0.2
 * Fix synchronization promise at widget initialization.