+++ sources/www/widgets/numeric-display/changelog.md
@@ -1,5 +1,6 @@
-### 1.0.1
-* Fix minimum value for the display precision.
-
 ### 1.0.2
 * Fix display for values with no units.
+* Better size proportions/fix display with firefox.
+
+### 1.0.1
+* Fix minimum value for the display precision.